Upgrade hibernate from 3.4 to 3.6
This made the use of joda-time classes instead of java.util classes
possible. This will make date and time calulcation much more easier as it
would be with standard java library.

This also reduce the transfer costs between API and the webservice

Also adjust the database to fit with the values.